implement user login and permission logics
