+++ accounts/templates/accounts/login.html
@@ -13,7 +13,7 @@
     <h3 class="panel-title text-center">Log-in to mysite</h3>
   </div>
   <div class="panel-body">
-    <form action="{% url 'login' %}" method="post" role="form">
+    <form action="{% url 'accounts:login' %}" method="post" role="form">
       {% csrf_token %}
       <div class="input-group" style="margin-bottom:10px">
         <span class="input-group-addon"><i class="fa fa-user fa-fw"></i></span>
