+++ accounts/templates/accounts/logout.html
@@ -1,0 +1,15 @@
+{% extends "admin/base.html" %}
+
+{% block content %}
+
+<div class="panel panel-primary">
+  <div class="panel-heading">
+    <h3 class="panel-title text-center">Log-out</h3>
+  </div>
+  <div class="panel-body">
+    <div>ログアウトしました。</div>
+    <div><a href="{% url 'accounts:login' %}">Log-in again</a></div>
+  </div>
+</div>
+{% endblock %}
+
