+++ accounts/urls.py
@@ -1,12 +1,14 @@
 from django.urls import path
-from django.contrib.auth.views import login,logout
+from django.contrib.auth.views import login as django_login
 from . import views
 
+app_name = 'accounts'
 urlpatterns = [
     path('', views.index, name='index'),
-    path('login/', login,
+    path('login/', django_login,
         {'template_name': 'accounts/login.html'},
         name='login'),
-    path('logout/', logout, name='logout'),
+    path('redirect/', views.login_redirect, name='loginredirect'),
+    path('logout/', views.logout, name='logout'),
 ]
