+++ accup/settings.py
@@ -115,3 +115,7 @@ USE_TZ = True
 STATIC_URL = '/static/'
 STATIC_ROOT = os.path.join(BASE_DIR, 'static')
 
+
+# redirect setting for after logined
+LOGIN_REDIRECT_URL = 'accounts:loginredirect'
+
